home sản phẩm

+++ app/src/main/java/com/example/hairsalon/activity/order/PayActivity.java
@@ -1,15 +1,22 @@
-package com.example.hairsalon.activity.pay;
+package com.example.hairsalon.activity.order;
 
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.DialogInterface;
+import android.graphics.Bitmap;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
 import android.widget.ArrayAdapter;
+import android.widget.ImageView;
+import android.widget.LinearLayout;
 import android.widget.Toast;
 
 import com.android.volley.Request;
+import com.android.volley.RequestQueue;
 import com.android.volley.VolleyError;
+import com.android.volley.toolbox.ImageRequest;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
 import com.example.hairsalon.adapter.CartItemAdapter;
@@ -23,8 +30,10 @@ import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
+import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.Date;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -33,38 +42,8 @@ import retrofit2.Callback;
 import com.android.volley.Response;
 
 import android.content.Intent;
-import android.graphics.Bitmap;
-import android.os.Bundle;
-import android.util.Log;
-import android.view.View;
-import android.widget.ImageView;
-import android.widget.Toast;
 
-import com.android.volley.Request;
-import com.android.volley.RequestQueue;
-import com.android.volley.Response;
-import com.android.volley.VolleyError;
-import com.android.volley.toolbox.ImageRequest;
-import com.android.volley.toolbox.StringRequest;
-import com.android.volley.toolbox.Volley;
-import com.example.hairsalon.R;
-import com.example.hairsalon.activity.cart.CartActivity;
-import com.example.hairsalon.adapter.GridAdapter;
-import com.example.hairsalon.api.ApiService;
-import com.example.hairsalon.constants.Constant;
-import com.example.hairsalon.databinding.ActivityDetailProductBinding;
-import com.example.hairsalon.model.ResponseData;
-
-import org.json.JSONException;
-import org.json.JSONObject;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
 import java.util.Objects;
-
-import retrofit2.Call;
 
 
 public class PayActivity extends AppCompatActivity {
@@ -75,19 +54,63 @@ public class PayActivity extends AppCompatActivity {
     ArrayList<Map<String, Object>> cartItemList = new ArrayList<>();
 
     ArrayList<CartItem> dataArrayList = new ArrayList<>();
-
+    Double price = 0.0;
+    Integer productItemId;
     private Double totalPrice;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         binding = ActivityPayBinding.inflate(getLayoutInflater());
         setContentView(binding.getRoot());
-        getCartItems();
+        Intent intent = getIntent();
+        if (intent.getStringExtra("detailName") != null) {
+            binding.linearLayout.setVisibility(View.VISIBLE);
+            binding.listViewProducts.setVisibility(View.GONE);
+            String name = intent.getStringExtra("detailName");
+            price = intent.getDoubleExtra("detailPrice", 0);
+            String imageUrl = intent.getStringExtra("imageUrl");
+            productItemId = intent.getIntExtra("productItemId", 0);
+            binding.productName.setText(name);
+            binding.productPrice.setText(String.valueOf(price));
+            binding.productQuantity.setText("1");
+            binding.xSeparator.setVisibility(View.VISIBLE);
+            binding.productQuantity.setVisibility(View.VISIBLE);
+            binding.totalPrice.setText(String.valueOf(price));
+            RequestQueue requestQueue = Volley.newRequestQueue(PayActivity.this);
+            ImageRequest imageRequest = new ImageRequest(
+                    imageUrl,
+                    new Response.Listener<Bitmap>() {
+                        @Override
+                        public void onResponse(Bitmap response) {
+                            // Gán hình ảnh vào ImageView
+                            binding.productImage.setImageBitmap(response);
+                        }
+                    },
+                    0,
+                    0,
+                    ImageView.ScaleType.CENTER_INSIDE,
+                    Bitmap.Config.RGB_565,
+                    new Response.ErrorListener() {
+                        @Override
+                        public void onErrorResponse(VolleyError error) {
+                            // Xử lý khi có lỗi xảy ra
+                            Log.e("Volley", "Failed to load image", error);
+                        }
+                    }
+            );
+            requestQueue.add(imageRequest);
+
+        } else {
+            binding.linearLayout.setVisibility(View.GONE);
+            binding.listViewProducts.setVisibility(View.VISIBLE);
+            getCartItems();
+        }
+
         setupSpinners();
         binding.payButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                createOrder();
+                showConfirmationDialog();
             }
         });
     }
@@ -102,17 +125,19 @@ public class PayActivity extends AppCompatActivity {
                         cartItemList = (ArrayList<Map<String, Object>>) responseData.getData();
                         totalPrice = 0.0;
                         for (Map<String, Object> cartItem : cartItemList) {
-                            totalPrice +=  (double) cartItem.get("quantity") *  (double) cartItem.get("price");
+
                             Integer id = ((Number) cartItem.get("id")).intValue();
                             String productItemName = (String) cartItem.get("productItemName");
-                            double price = (double) cartItem.get("price");
                             Integer quantity = ((Number) Objects.requireNonNull(cartItem.get("quantity"))).intValue();
                             String imageUrl = (String) cartItem.get("imageUrl");
-                            totalPrice += quantity * price;
-                            CartItem cartItemAdded = new CartItem(id, productItemName, quantity, imageUrl, price);
+                            double priceProduct = (double) cartItem.get("price");
+                            totalPrice +=  quantity * priceProduct;
+
+                            CartItem cartItemAdded = new CartItem(id, productItemName, quantity, imageUrl, priceProduct);
                             dataArrayList.add(cartItemAdded);
                         }
                         binding.totalPrice.setText(String.valueOf(totalPrice));
+
                         Collections.reverse(dataArrayList);
                         cartItemAdapter = new CartItemAdapter(PayActivity.this, dataArrayList);
                         Log.i("productItemList", dataArrayList.toString());
@@ -130,10 +155,6 @@ public class PayActivity extends AppCompatActivity {
                 Log.e("Error", "API call failed: " + t.getMessage());
             }
         });
-    }
-
-    private void showCartItems() {
-
     }
 
     private void setupSpinners() {
@@ -156,22 +177,38 @@ public class PayActivity extends AppCompatActivity {
             JSONObject requestBody = new JSONObject();
             requestBody.put("customerId", 1);
             requestBody.put("payId", 1);
-            requestBody.put("totalPrice", totalPrice);
-            requestBody.put("orderDate", "2023-10-10");
+            SimpleDateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
+            String currentDate = dateFormat.format(new Date());
+            requestBody.put("orderDate", currentDate);
             JSONArray orderItemListJson = new JSONArray();
-            for (Map<String, Object> cartItem : cartItemList) {
+            if (price != 0.0) {
+                requestBody.put("totalPrice", price);
                 JSONObject orderItemJson = new JSONObject();
-                orderItemJson.put("productItemId", cartItem.get("productItemId"));
-                orderItemJson.put("price", cartItem.get("price"));
-                orderItemJson.put("quantity", cartItem.get("quantity"));
+                orderItemJson.put("productItemId", productItemId);
+                orderItemJson.put("price", price);
+                orderItemJson.put("quantity", 1);
                 orderItemListJson.put(orderItemJson);
+                requestBody.put("orderItemList", orderItemListJson);
             }
-            requestBody.put("orderItemList", orderItemListJson);
+            else {
+                for (Map<String, Object> cartItem : cartItemList) {
+                    JSONObject orderItemJson = new JSONObject();
+                    orderItemJson.put("productItemId", cartItem.get("productItemId"));
+                    orderItemJson.put("price", cartItem.get("price"));
+                    orderItemJson.put("quantity", cartItem.get("quantity"));
+                    orderItemListJson.put(orderItemJson);
+                }
+                requestBody.put("totalPrice", totalPrice);
+                requestBody.put("orderItemList", orderItemListJson);
+                deleteAllCartItems();
+            }
+
             StringRequest request = new StringRequest(Request.Method.POST, apiUrl,
                     new Response.Listener<String>() {
                         @Override
                         public void onResponse(String response) {
-                            deleteAllCartItems();
+                            Intent intent = new Intent(PayActivity.this, SuccessPay.class);
+                            startActivity(intent);
                             Toast.makeText(getApplicationContext(), "Đơn hàng của bạn đã được thanh toán", Toast.LENGTH_SHORT).show();
                         }
                     },
@@ -205,13 +242,11 @@ public class PayActivity extends AppCompatActivity {
             @Override
             public void onResponse(Call<Void> call, retrofit2.Response<Void> response) {
                 if (response.isSuccessful()) {
-                    Intent intent = new Intent(PayActivity.this, SuccessPay.class);
-                    startActivity(intent);
+
                 } else {
                     Log.e("Error", "API call failed with error code: " + response.code());
                 }
             }
-
             @Override
             public void onFailure(Call<Void> call, Throwable t) {
                 Log.e("Error", "API call failed: " + t.getMessage());
@@ -219,6 +254,23 @@ public class PayActivity extends AppCompatActivity {
         });
     }
 
-
+    private void showConfirmationDialog() {
+        AlertDialog.Builder builder = new AlertDialog.Builder(this);
+        builder.setTitle("Xác nhận đặt hàng");
+        builder.setMessage("Xác nhận đặt đơn hàng này?");
+        builder.setPositiveButton("Xác nhận", new DialogInterface.OnClickListener() {
+            @Override
+            public void onClick(DialogInterface dialog, int which) {
+                createOrder();
+            }
+        });
+        builder.setNegativeButton("Hủy", new DialogInterface.OnClickListener() {
+            @Override
+            public void onClick(DialogInterface dialog, int which) {
+                dialog.dismiss();
+            }
+        });
+        builder.show();
+    }
 
 }
